Changed how the database worked and added mock data

+++ app/src/main/java/com/logger/car/androidcarmaintenanceapp/domain/GasLogEntry.kt
@@ -5,5 +5,5 @@ import java.util.*
 data class GasLogEntry(
         override val entryDate: Date,
         override val mileage: Int,
-        val gallonsAdded: Int
+        val gallonsAdded: Double
 ): BaseLogEntry
